feat: usage visibility for customers and operators
Both consoles showed money moving, not work done. The settled
payment_session rows already carry capability, offering, key, units,
funded and billed, so the usage domain now reads them directly:
/v1/accounts/me/usage/{overview,jobs,summary} for customers (session
cookie or API key) and /v1/admin/usage/{summary,jobs,attention} plus
per-user views for operators. Overview reports available, held and spent
credit against the billing window; attention lists open jobs still
holding funds past a stale threshold and settlement verification
failures, which jobs settle and sessions close now record as
server.settlement_verification_failed events.

Every wei amount on the wire is an integer string via the shared
WeiDecimal serializer (balance, ledger, payments and admin views used
to emit NUMERIC exponents such as 1.20E+14). The never-written
usage_record table is dropped (migration 0024); one source of truth.

Portal: dashboard shows Available / Held / Spent this period with the
cap meter and a 30-day sparkline, a Usage page with filters, paging and
CSV export, spend per API key, and ledger reasons in user language.
Admin: attention counters and a 7-day fleet chart on the overview,
spent/held columns and a per-user usage drawer on Users, and a fleet
Usage page. Product spec 006 written; OpenAPI snapshot and generated
SDK types refreshed. Beads epic loc-999; follow-up loc-gcg for the
SDK-facing int wei fields.

+++ src/livepeer_open_clearinghouse/domains/admin/runtime.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import uuid
+from decimal import Decimal
 from typing import Any
 
 from fastapi import APIRouter, HTTPException, Response, status
@@ -82,7 +83,7 @@ async def list_all_users_endpoint(
                 email=u.email,
                 email_verified_at=u.email_verified_at,
                 approved=approved,
-                balance_wei=balance_wei,
+                balance_wei=Decimal(balance_wei),
                 created_at=u.created_at,
             )
             for (u, approved, balance_wei) in rows
@@ -230,8 +231,8 @@ async def list_deposit_snapshots_endpoint(
             DepositSnapshotView(
                 id=r.id,
                 taken_at=r.taken_at,
-                deposit_wei=int(r.deposit_wei),
-                reserve_wei=int(r.reserve_wei),
+                deposit_wei=Decimal(r.deposit_wei),
+                reserve_wei=Decimal(r.reserve_wei),
                 withdraw_round=r.withdraw_round,
                 current_round=r.current_round,
                 ticket_validity_period=r.ticket_validity_period,
